Update README.md
This repository provides a comprehensive guide to WebGL, the JavaScript 
API for creating 2D and 3D graphics on the web. You will learn the theory 
and practice of WebGL, such as how to use shaders, buffers, textures, 
matrices, cameras, and events. 
You will learn the essentials of WebGL, such as how to create triangles, 
colors, textures, matrices, and events. You will also learn how to use 
JavaScript and HTML5 to manipulate the graphics and interact with the 
user. You will apply your skills to create various projects, such as a star, a 
cube, a sphere, etc. You will also learn how to use loops, functions, classes, 
and dropdown menus to make your code more efficient and modular. This 
repository is suitable for anyone who wants to learn how to use WebGL in a 
creative and experimental way